Allow TrustAnchors to enforce a custom policy
This policy is expressed as a function, which receives
the end-entity certificate, subject, spki and validity times.
If this function returns any error, that trust anchor is not
used for chain creation.

This involves Cert keeping a reference to the original DER
encoding of the certificate.  Having the whole end-entity
certificate is useful in the policy to implement a whitelist
(for example, Mozilla's CNNIC whitelist consists of SHA256
hashes of the whole certificates.)

It also involves manually implementing fmt::Debug for
TrustAnchor because functions obviously can't be automatically
derived for this.